+++ NEWS.md
@@ -1,3 +1,10 @@
+# insight (devel)
+
+## Changes
+
+* `find_response()`, `find_predictors()`, and `find_variables()` now also work
+  for objects from package *marginaleffects*.
+
 # insight 1.5.2
 
 ## Changes